Threads system replaced by a timers.
The architecture based on multiple fibers that communicates through messages is replaced by a sequential architecture controled by timers of the GUI.
This decision was made due to shared variables that are too much intrusive and prevent the architecture to be independent from the parallelisme method.
Casting shared variables to local ones cause non-deterministic crashes since the compiler generate bad optimisations and is now forbidden.